+++ changelog.md
@@ -2,4 +2,5 @@
 
 ## Version 0.2.0
 
+- **BREAKING CHANGE:** `TprHeader` field `has_coordinates` has been renamed to `has_positions`.
 - `minitpr` can now parse intermolecular bonds. Intermolecular bonds are added to the end of the `bonds` vector in `TprFile::TprTopology`.